Implement a versioning system
Allowing users to archive a version of a website can be useful for
education websites that want to preserve past iterations of a course,
and is key for documentation websites to preserve past documentation
versions.

Let’s implement said feature, allowing authors to archive the current
state of the site using the `markbind archive <version_name>` command.

Authors might also want perform maintenance on past versions in the
future, such as renaming the baseUrl of the site.
Additionally, some versions of markbind may be prone to bugs that can be
rectified with newer versions of markbind.
Let’s also archive the source files of the current state hence,
allowing said maintenance or allowing authors to fix past versions
of the site with newer markbind versions.
Let’s also provide the `markbind rebuild-archive` command that
automatically rebuilds all past versions of the site.